+++ share/ci/scripts/multi/install_imath.sh
@@ -6,30 +6,49 @@ set -ex
 
 IMATH_VERSION="$1"
 INSTALL_TARGET="$2"
+remove=1
 
-git clone https://github.com/AcademySoftwareFoundation/Imath.git
+if [ $3 == "keep" ]; then
+    remove=0
+fi
+
+if [ ! -d "Imath" ]; then
+    git clone https://github.com/AcademySoftwareFoundation/Imath.git
+else
+    cd Imath
+    git reset --hard origin/main
+    git checkout main
+    git fetch
+    git pull
+    cd ..
+fi
+
 cd Imath
 
 if [ "$IMATH_VERSION" == "latest" ]; then
-    git checkout release
     LATEST_TAG=$(git describe --abbrev=0 --tags)
+    git branch -D ${LATEST_TAG} >/dev/null 2>&1 || true
     git checkout tags/${LATEST_TAG} -b ${LATEST_TAG}
 else
+    git branch -D v${IMATH_VERSION} >/dev/null 2>&1 || true
     git checkout tags/v${IMATH_VERSION} -b v${IMATH_VERSION}
 fi
 
-mkdir build
+mkdir -p build
 cd build
 cmake -DCMAKE_BUILD_TYPE=Release \
-      ${INSTALL_TARGET:+"-DCMAKE_INSTALL_PREFIX="${INSTALL_TARGET}""} \
-      -DBUILD_TESTING=OFF \
-      -DPYTHON=OFF \
-      -DCMAKE_POSITION_INDEPENDENT_CODE=ON \
-      ../.
+    ${INSTALL_TARGET:+"-DCMAKE_INSTALL_PREFIX="${INSTALL_TARGET}""} \
+    -DBUILD_TESTING=OFF \
+    -DPYTHON=OFF \
+    -DCMAKE_OSX_SYSROOT=/Library/Developer/CommandLineTools/SDKs/MacOSX.sdk \
+    -DCMAKE_POSITION_INDEPENDENT_CODE=ON \
+    ../.
 cmake --build . \
-      --target install \
-      --config Release \
-      --parallel 2
+    --target install \
+    --config Release
 
 cd ../..
-rm -rf Imath
+
+if [ $remove == 1 ]; then
+    rm -rf Imath
+fi
